Update class.Towerville2056.rb
Method_Missing saga continues.  This seems to be a dead-end, unless I want to monkey-patch INT, STR and HASH.  So, will wind up back-tracking a bit to implement this.

Essentially, "method of property of class" isn't a thing.  So, switching back to rails-esque path again, but keeping method_missing as tech demo.